% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebsummary.R
\name{ebsummary}
\alias{ebsummary}
\title{ebsummary Generic}
\usage{
ebsummary(x)
}
\arguments{
\item{x}{An object.}
}
\description{
This function is used as a generic for the \code{ebsummary} object class. This class is
    used to store output from the method \code{\link{summary.eb}}, which summariezes
    output from the models \code{\link{EpiBayes_s}} and \code{\link{EpiBayes_ns}}.
}
\details{
The \code{\link{EpiBayes_s}} and \code{\link{EpiBayes_ns}} functions return objects of the
    class \code{eb}, for which this is the generic. This class of objects is used to
    store the output of the EpiBayes models and to reference the \code{print},
    \code{summary}, and \code{plot} methods for these output objects.
}
\seealso{
Other classes' generics used within this package include \code{\link{ebhistorical}} for
    output objects of the \code{\link{EpiBayesHistorical}} function and
    \code{\link{ebsamplesize}} for output objects of the
    \code{\link{EpiBayesSampleSize}} function.
}

