% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.eb.R
\name{plot.eb}
\alias{plot.eb}
\title{Plot Method for EpiBayes Output}
\usage{
\method{plot}{eb}(x, burnin = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{eb} (e.g., the output of functions
\code{\link{EpiBayes_ns}} or \code{link{EpiBayes_s}}).}

\item{burnin}{Number of MCMC iterations to discard from the beginning of the chain.
Integer scalar.}

\item{...}{Additional arguments to be passed to \code{plot}.}
}
\value{
One plotting window for each subzone including posterior distributions for the
    cluster-level prevalence across all time periods.
}
\description{
This method plots the output of the function \code{\link{EpiBayes_ns}} and
    \code{\link{EpiBayes_s}}. It does so by plotting the posterior distribution(s) of the
    cluster-level prevalence(s).
}

