% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.ebhistorical.R
\name{plot.ebhistorical}
\alias{plot.ebhistorical}
\title{Plot Method for EpiBayesHistorical Output}
\usage{
\method{plot}{ebhistorical}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ebhistorical} (e.g., the output of
function \code{\link{EpiBayesHistorical}}).}

\item{...}{Additional arguments to be passed to \code{plot}.}
}
\value{
One plotting window for each subzone including posterior distributions for the
    cluster-level prevalence across all time periods.
}
\description{
This method plots the output of the function \code{\link{EpiBayesHistorical}}. It does
    so by plotting the posterior distributions of the cluster-level prevalence across
    all time periods in different colors within a particular subzone. Similar plots
    are made for each subzone in separate plotting windows (\emph{Caveat: too many
    subzones can make manipulation of the plotting windows unmageable}).
}

