% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.ebhistoricalsummary.R
\name{plot.ebhistoricalsummary}
\alias{plot.ebhistoricalsummary}
\title{Plot Method for Summary of EpiBayes Historical Object}
\usage{
\method{plot}{ebhistoricalsummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ebhistoricalsummary} (e.g., the output of
method \code{\link{summary.ebhistorical}}).}

\item{...}{Additional arguments to be passed on to plot.}
}
\value{
The summary statistics are plotted in the current plotting window. The colors and
    plotting characters represent the various subzones and the actual points in the plot
    are the summary statistics (vertical axis) measured in a particular time period
    (horizontal axis).
}
\description{
This function plots summary measurements for posterior distributions of cluster-level
    prevalences across all time periods considered. It does so by plotting the output
    from the \code{\link{summary.ebhistorical}} method, which is of class
    \code{ebhistoricalsummary}.
}
\seealso{
This is a method for objects of class \code{ebhistoricalsummary} returned by the method
    \code{\link{summary.ebhistorical}} and creates its own class of object much like the
    summary method for \code{lm} objects.
}

