% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.eb.R
\name{print.eb}
\alias{print.eb}
\title{Print Method for EpiBayes Object}
\usage{
\method{print}{eb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eb} (e.g., the output of
function \code{\link{EpiBayes_ns}}) or \code{\link{EpiBayes_s}}).}

\item{...}{Additional arguments to be passed on to summary.}
}
\value{
The summary statistics are returned in a list with the first entry containing the
    simulation output (\code{p2.tilde}, \code{p4.tilde}, and \code{p6.tilde}), the next
    containing summary measures for the first ten replicated data sets' \code{gam}, and
    the rest containing summary measures for the first ten replicated data sets'
    \code{tau} values (one for each subzone, if applicable).
    The summary measurements taken on the posterior distributions include the posterior
    mean, standard deviation, standard error of the mean, time-series adjusted standard
    error of the mean, and the lower and upper HPD interval limits, in that order.
    For reference purposes, below are the descriptions for the summarized variables.
    By default, the function returns summary values for up to ten replicated data sets of
    \code{gam} and \code{tau} and reports 95% HPD intervals.

\tabular{lll}{
    Output \tab Description \cr
    \code{p2.tilde} \tab Proportion of simulated data sets that result in the probability of \code{poi} prevalence \emph{below} \code{poi.thresh} with probability \code{p1} \cr
    \code{p4.tilde} \tab Proportion of simulated data sets that result in the probability of \code{poi} prevalence \emph{above} \code{poi.thresh} with probability \code{p1} \cr
    \code{p6.tilde} \tab Proportion of simulated data sets that result in the probability of \code{poi} prevalence \emph{between} \code{poi.lb} and \code{poi.ub} with probability \code{p1} \cr
    \code{taumat} \tab Posterior distributions of the cluster-level prevalence for all simulated data sets (i.e., \code{reps}) \cr
    \code{gammat} \tab Posterior distribution of the subzone-level prevalence (3-level) OR Posterior distribution of the probability of the disease being in the region (2-level) \cr
}
}
\description{
This function prints basic summary measurements for posterior distributions of cluster-
    level prevalences across all time periods considered. It does so by examining the
    object output by the \code{\link{EpiBayes_ns}} or \code{\link{EpiBayes_s}} function
    of class \code{eb}.
}
\seealso{
This is a method for objects of class \code{eb} returned by the function
    \code{\link{EpiBayes_ns}} or \code{\link{EpiBayes_s}}. The \code{summary} method for
    the \code{eb} object class allows for customization of this output.
}

