% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ebhistorical.R
\name{print.ebhistorical}
\alias{print.ebhistorical}
\title{Print Method for EpiBayes Historical Object}
\usage{
\method{print}{ebhistorical}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ebhistorical} (e.g., the output of
function \code{\link{EpiBayesHistorical}}).}

\item{...}{Additional arguments to be passed on to summary.}
}
\value{
The posterior means are returned in a matrix with rows indexed by time period and
    columns indexed by subzone. The output of this method can be customized using
    the \code{summary} method for this \code{ebhistorical} class of object.
}
\description{
This function gives basic summary statistics for posterior distributions of cluster-level
    prevalences across all time periods considered. It does so by examining the object
    output by the \code{\link{EpiBayesHistorical}} function of class
    \code{ebhistorical}.
}
\seealso{
This is a method for objects of class \code{ebhistorical} returned by the function
    \code{\link{EpiBayesHistorical}}. The \code{summary} method for
    the \code{ebhistorical} object class allows for customization of this output.
}

