% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ebhistoricalsummary.R
\name{print.ebhistoricalsummary}
\alias{print.ebhistoricalsummary}
\title{Print Method for EpiBayes Historical Object}
\usage{
\method{print}{ebhistoricalsummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ebhistoricalsummary} (e.g., the output of
function \code{\link{summary.ebhistorical}}).}

\item{...}{Additional arguments to be passed on to print.}
}
\value{
The summary statistics are returned in a matrix with rows indexed by time period and
    columns indexed by subzone. The body of the matrix is filled with the summary
    statistics requested for by the argument \code{sumstat}.
}
\description{
This function prints the output of objects of class \code{ebhistoricalsummary} from the
    method \code{\link{summary.ebhistorical}} in a nice form.
}
\seealso{
This is a method for objects of class \code{ebhistoricalsummary} returned by the method
    \code{\link{summary.ebhistorical}}, which creates its own class of object much like
    the summary method for \code{lm} objects. There is an additional plot method that
    will plot the summary output called \code{\link{plot.ebhistoricalsummary}}.
}

