% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ebsamplesize.R
\name{print.ebsamplesize}
\alias{print.ebsamplesize}
\title{Print Method for EpiBayes Sample Size Object}
\usage{
\method{print}{ebsamplesize}(x, out.ptilde = c("p2.tilde", "p4.tilde",
  "p6.tilde"), ...)
}
\arguments{
\item{x}{An object of class \code{ebsamplesize} (e.g., the output of
function \code{\link{EpiBayesSampleSize}}).}

\item{out.ptilde}{Indicators for the desired output of the printing. Can be one or any
output is desired, supply a character vector of length 1, if 2 are desired, supply
a character vector of length 2, and 3 for 3. Character vector.}

\item{...}{Additional arguments to be passed on to print.}
}
\value{
The summary statistics are returned in a matrix with rows indexed by time period and
    columns indexed by subzone. The body of the matrix is filled with the summary
    statistics requested for by the argument \code{sumstat}.
}
\description{
This function prints the output of objects of class \code{ebsamplesize} from the
    method \code{\link{EpiBayesSampleSize}} in a nice form.
}
\seealso{
This is a method for objects of class \code{ebhistoricalsummary} returned by the method
    \code{\link{summary.ebhistorical}}, which creates its own class of object much like
    the summary method for \code{lm} objects. There is an additional plot method that
    will plot the summary output called \code{\link{plot.ebhistoricalsummary}}.
}

